GunRotating
Rough gun movement logic, it might almost make more sense to have the gun be a seperate entity to the player, maybe even a child of the camera.

Also its a bit floaty, not sure why its like that but if we dont like it I will tighen it up.

Any questions ask Josh